study/udemy/automate-everything-with-python/requirements.txt
altair==4.2.0
appdirs==1.4.4
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
black==22.8.0
blinker==1.5
bs4==0.0.1
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
cssselect==1.1.0
cssutils==2.6.0
decorator==5.1.1
distro==1.7.0
entrypoints==0.4
et-xmlfile==1.1.0
fake-useragent==0.1.11
flake8==5.0.4
Flask==2.2.2
fpdf==1.7.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.11.1
google-auth-oauthlib==0.5.3
gspread==5.5.0
h11==0.13.0
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
lxml==4.9.1
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
numpy==1.23.3
oauthlib==3.2.1
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.5.0
parse==1.19.0
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
premailer==3.10.0
protobuf==3.20.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pydeck==0.8.0b3
pyee==8.2.2
pyflakes==2.5.0
Pygments==2.13.0
Pympler==1.0.1
PyMuPDF==1.20.2
pyparsing==3.0.9
pyppeteer==1.0.2
PyQt6==6.3.1
PyQt6-Qt6==6.3.2
PyQt6-sip==13.4.0
pyquery==1.4.3
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-html==0.10.0
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
selenium==4.4.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
streamlit==1.12.2
tabula-py==2.5.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
w3lib==2.0.1
watchdog==2.1.9
websockets==10.3
Werkzeug==2.2.2
wsproto==1.2.0
yagmail==0.15.293
zipp==3.8.1
